se modifica el main y se pasa a un nuevo archivo
